open-source-teamproject/requirements.txt
Django>=4.2
mysqlclient
djangorestframework